hotfix(core) run shms check in early init phase
Some shms, and especially the `kong` one might be used inside of the
global patches (or in any of the other modules we require for that is).

This bug was caught on the CI environment with OpenResty 1.11.2.1, which
calls `math.randomseed()` inside of its main chunk from the
lua-resty-dns library, which triggers our custom
`globalpatches.randomseed` behavior, which needs the `kong` shm.